sim tests: move helper functions into test_utils file
